reporting-section/requirements.txt
# Kafka
kafka-python==2.0.2

# LangChain & OpenAI
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.30.0

# Vector DB
chromadb==0.4.22

# BM25
rank-bm25==0.2.2

# 텍스트 처리
tiktoken==0.5.2
nltk==3.8.1

# 데이터베이스
psycopg2-binary==2.9.9

# 유틸리티
python-dotenv==1.0.0
pydantic==2.5.3
pydantic_settings
sqlalchemy
pymysql
openai